feature: 동아리 기본정보 수정 이미지 null 확인로직 추가
