+++ src/main/java/kr/hanjari/backend/domain/Club.java
@@ -5,7 +5,7 @@ import kr.hanjari.backend.domain.common.BaseEntity;
 import kr.hanjari.backend.domain.enums.ClubCategory;
 import kr.hanjari.backend.domain.enums.RecruitmentStatus;
 import kr.hanjari.backend.web.dto.club.request.ClubDetailRequestDTO;
-import kr.hanjari.backend.web.dto.club.request.CommonClubDTO;
+import kr.hanjari.backend.web.dto.club.request.ClubBasicInformationDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -75,7 +75,6 @@ public class Club extends BaseEntity {
     public void updateClubDetails(ClubDetailRequestDTO detail) {
         this.recruitmentStatus = detail.recruitmentStatus();
         this.leaderName = detail.leaderName();
-
         this.leaderPhone = detail.leaderPhone();
         this.membershipFee = detail. membershipFee();
         this.meetingSchedule = detail.activities();
@@ -83,7 +82,7 @@ public class Club extends BaseEntity {
         this.applicationUrl = detail.applicationUrl();
     }
 
-    public void updateClubCommonInfo(CommonClubDTO commonInfo) {
+    public void updateClubCommonInfo(ClubBasicInformationDTO commonInfo) {
         this.name = commonInfo.clubName();
         this.leaderEmail = commonInfo.leaderEmail();
         this.category = ClubCategory.valueOf(commonInfo.category());
