+++ src/main/java/kr/hanjari/backend/web/dto/club/request/ClubBasicInformationDTO.java
@@ -1,0 +1,27 @@
+package kr.hanjari.backend.web.dto.club.request;
+
+import kr.hanjari.backend.domain.ClubRegistration;
+import kr.hanjari.backend.domain.enums.ClubCategory;
+
+
+public record ClubBasicInformationDTO(String clubName,
+                                      String leaderEmail,
+                                      String category,
+                                      String oneLiner,
+                                      String briefIntroduction
+) {
+
+    public static ClubBasicInformationDTO of(String clubName, String leaderEmail, String category, String oneLiner, String briefIntroduction) {
+        return new ClubBasicInformationDTO(clubName, leaderEmail, category, oneLiner, briefIntroduction);
+    }
+
+    public ClubRegistration toClubRegistration() {
+        return ClubRegistration.builder()
+                .name(clubName())
+                .category(ClubCategory.valueOf(category()))
+                .leaderEmail(leaderEmail())
+                .oneLiner(oneLiner())
+                .briefIntroduction(briefIntroduction())
+                .build();
+    }
+}
